requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.4.0=pypi_0
astor=0.8.1=pyh9f0ad1d_0
autograd=1.5=pyhd8ed1ab_0
autograd-gamma=0.5.0=pyh9f0ad1d_0
blas=1.0=mkl
bottleneck=1.3.5=py37h7deecbd_0
ca-certificates=2022.9.24=ha878542_0
cached_property=1.5.2=pyha770c72_1
cachetools=5.3.1=pypi_0
certifi=2022.9.24=pyhd8ed1ab_0
charset-normalizer=3.1.0=pypi_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.11.0=pyhd8ed1ab_0
fftw=3.3.9=h27cfd23_1
formulaic=0.5.2=pyhd8ed1ab_0
freetype=2.10.4=h0708190_1
future=0.18.2=py37h89c1867_5
google-auth=2.19.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
graphlib-backport=1.0.3=pyhd8ed1ab_0
grpcio=1.54.2=pypi_0
h5py=2.9.0=pypi_0
icu=67.1=he1b5a44_0
idna=3.4=pypi_0
importlib-metadata=5.0.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
interface_meta=1.3.0=pyhd8ed1ab_0
jinja2=3.1.2=pypi_0
joblib=1.1.0=pyhd3eb1b0_0
kiwisolver=1.4.2=py37h295c915_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libpng=1.6.37=hbc83047_0
libstdcxx-ng=11.2.0=h1234567_1
lifelines=0.27.1=pyhd8ed1ab_0
llvmlite=0.39.1=pypi_0
markdown=3.4.3=pypi_0
markupsafe=2.1.2=pypi_0
matplotlib-base=3.2.2=py37h1d35a4c_1
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py37h7f8727e_0
mkl_fft=1.3.1=py37hd3c417c_0
mkl_random=1.2.2=py37h51133e4_0
ncurses=6.3=h5eee18b_3
numba=0.56.4=pypi_0
numexpr=2.8.3=py37h807cd23_0
numpy=1.21.5=py37h6c91a56_3
numpy-base=1.21.5=py37ha15fc14_3
oauthlib=3.2.2=pypi_0
openssl=1.1.1q=h7f8727e_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.3.5=py37h8c16a72_0
pip=22.2.2=py37h06a4308_0
prettytable=3.4.1=pypi_0
protobuf=3.20.3=pypi_0
psutil=5.9.5=pypi_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pynndescent=0.5.8=pypi_0
pyparsing=3.0.9=py37h06a4308_0
python=3.7.13=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.7=2_cp37m
pytorch=1.12.0=py3.7_cuda11.3_cudnn8.3.2_0
pytorch-mutex=1.0=cuda
pytz=2022.1=py37h06a4308_0
readline=8.1.2=h7f8727e_1
requests=2.31.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
scikit-learn=1.0.2=py37h51133e4_1
scipy=1.7.3=py37h6c91a56_2
setuptools=63.4.1=py37h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.38.2=hc218d9a_0
tensorboard=2.11.2=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.11=h1ccaba5_0
torch-geometric=2.3.1=pypi_0
torch-scatter=2.0.9=pypi_0
torch-sparse=0.6.15=pypi_0
tornado=6.1=py37h540881e_3
tqdm=4.64.1=pypi_0
typing-extensions=4.3.0=py37h06a4308_0
typing_extensions=4.3.0=py37h06a4308_0
umap-learn=0.5.3=pypi_0
urllib3=1.26.16=pypi_0
wcwidth=0.2.5=pypi_0
werkzeug=2.2.3=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
wrapt=1.14.1=py37h540881e_0
xz=5.2.6=h5eee18b_0
zipp=3.9.0=pypi_0
zlib=1.2.11=0
